Redirection error
Redirection error in tables with rows smaller than three. For example when entering a table and searching for a field that results in three rows when ordering it by id or by date phpPgAdmin redirected me to public schema. The solution is to remove the redirect to tables.php and change it to display.php in the $ actions variable in the multiactions index, sub index url in line 770.